+++ ouds_core/lib/l10n/gen/ouds_localizations_en.dart
@@ -118,7 +118,7 @@ class OudsLocalizationsEn extends OudsLocalizations {
 
   @override
   String core_pinCodeInput_digitCode_label_a11y(Object current) {
-    return 'Digit code $current';
+    return '$current digit';
   }
 
   @override
@@ -131,6 +131,11 @@ class OudsLocalizationsEn extends OudsLocalizations {
 
   @override
   String get core_pinCodeInput_error_a11y => 'Error: Invalid code';
+
+  @override
+  String core_pinCodeInput_digitPosition_a11y(Object current, Object total) {
+    return 'Digit $current of $total';
+  }
 
   @override
   String get core_topAppBar_backNavigationIcon_a11y => 'Back';