+++ ouds_theme_orange/lib/components/orange_alert_tokens.dart
@@ -33,7 +33,8 @@ class OrangeAlertTokens extends OudsAlertTokens {
   @override
   double get sizeMinHeight => providersTokens.sizeTokens.minInteractiveArea;
   @override
-  double get sizeMinHeightBottomActionPlacement => DimensionRawTokens.dimension1250;
+  double get sizeMinHeightBottomActionPlacement =>
+      DimensionRawTokens.dimension1250;
   @override
   double get sizeMinWidth => DimensionRawTokens.dimension2000;
   @override
@@ -41,9 +42,11 @@ class OrangeAlertTokens extends OudsAlertTokens {
   @override
   double get spaceColumnGapAction => providersTokens.spaceTokens.columnGapSmall;
   @override
-  double get spacePaddingBlock => providersTokens.spaceTokens.paddingBlockMedium;
+  double get spacePaddingBlock =>
+      providersTokens.spaceTokens.paddingBlockMedium;
   @override
-  double get spacePaddingInline => providersTokens.spaceTokens.paddingInlineLarge;
+  double get spacePaddingInline =>
+      providersTokens.spaceTokens.paddingInlineLarge;
   @override
   double get spaceRowGap => providersTokens.spaceTokens.rowGapSmall;
   @override